2019-01-11 update

+++ admin/category_update.php
@@ -1,35 +1,40 @@
 <?php
 include_once ("./_common.php");
 
-@mkdir(G5_DATA_PATH.'/cate/', G5_DIR_PERMISSION);
-@chmod(G5_DATA_PATH.'/cate/', G5_DIR_PERMISSION);
+if(!$parent_ca_id) {
 
-$upload_path = G5_DATA_PATH.'/cate/';
+    @mkdir(G5_DATA_PATH . '/cate/', G5_DIR_PERMISSION);
+    @chmod(G5_DATA_PATH . '/cate/', G5_DIR_PERMISSION);
 
-$sql = "select * from `categorys` where ca_id = '{$ca_id}'";
-$file = sql_fetch($sql);
+    $upload_path = G5_DATA_PATH . '/cate/';
 
-$icon = "";
+    $sql = "select * from `categorys` where ca_id = '{$ca_id}'";
+    $file = sql_fetch($sql);
 
-$error = "";
-if ($_FILES["icon"]["tmp_name"]) {
-    unlink($upload_path . $file["icon"]);
+    $icon = "";
 
-    $icon = $_FILES["icon"]["name"];
+    $error = "";
+    if ($_FILES["icon"]["tmp_name"]) {
+        unlink($upload_path . $file["icon"]);
 
-    $filetmp = explode(".", $icon);
+        $icon = $_FILES["icon"]["name"];
 
-    $filename = $filetmp[0] . "_" . date("Ymd_hms") . "." . $filetmp[1];
+        $filetmp = explode(".", $icon);
 
-    $uploadfile = $upload_path . $filename;
+        $filename = $filetmp[0] . "_" . date("Ymd_hms") . "." . $filetmp[1];
 
-    $error .= move_uploaded_file($_FILES["icon"]["tmp_name"], $uploadfile);
+        $uploadfile = $upload_path . $filename;
 
-    $where = " , icon = '{$filename}'";
+        $error .= move_uploaded_file($_FILES["icon"]["tmp_name"], $uploadfile);
+
+        $where = " , icon = '{$filename}'";
+    }
+
+
+    $sql = "update `categorys` set cate_name = '{$cate_name}', cate_tag = '{$cate_tag}', info_text1 = '{$info_text1}', info_text2 = '{$info_text2}' {$where} WHERE ca_id = '{$ca_id}'";
+}else{
+    $sql = "update `categorys` set cate_name = '{$cate_name}', cate_tag = '{$cate_tag}', info_text1 = '{$info_text1}', info_text2 = '{$info_text2}' WHERE ca_id = '{$ca_id}'";
 }
-
-
-$sql = "update `categorys` set cate_name = '{$cate_name}', cate_tag = '{$cate_tag}', info_text1 = '{$info_text1}', info_text2 = '{$info_text2}' {$where} WHERE ca_id = '{$ca_id}'";
 if(sql_query($sql)){
     alert("카테고리가 수정되었습니다.");
 }else{
